Fix compute

+++ lib/Match.cpp
@@ -56,7 +56,7 @@ template <class R, class F, class P> void Match<R, F, P>::compute(R& result, con
                 // Similarity values are scaled for integers over [0,1000], for render % is fine...
                 m_pairs.emplace_back(&p.probe()->minutiae()[i], &p.candidate()->minutiae()[i], std::lround(static_cast<float>(p.similarity()) / 10.0f));
             }
-            if constexpr (std::is_same_v<R, uint16_t>) {
+            if constexpr (std::is_same_v<R, uint8_t>) {
                 m_pairs.emplace_back(&p.probe()->minutiae()[i], &p.candidate()->minutiae()[i]);
             }
         }
@@ -135,7 +135,7 @@ template <class R, class F, class P> void Match<R, F, P>::compute(R& result, con
         }
         maxMatched = std::max(maxMatched, matched);
     }
-    if constexpr (std::is_same_v<R, uint16_t>) {
+    if constexpr (std::is_same_v<R, uint8_t>) {
         if (maxMatched > param.MinimumMinutiae) {
             result = static_cast<uint8_t>((maxMatched * maxMatched * 100) / (probe.minutiaeCount() * candidate.minutiaeCount()));
         }
